test: consolidate GLM-5 model coverage
